upgrade redukt gradle and libs
