+++ README.md
@@ -1,6 +1,6 @@
-[![Build Status](https://travis-ci.org/raulccabreu/redukt.svg?branch=master)](https://travis-ci.org/raulccabreu/redukt)
-[![](https://jitpack.io/v/raulccabreu/redukt.svg)](https://jitpack.io/#raulccabreu/redukt)
-[![codebeat badge](https://codebeat.co/badges/50fb8d27-6eca-424e-9bbe-6f469b95cec9)](https://codebeat.co/projects/github-com-raulccabreu-redukt-master)
+[![Build Status](https://travis-ci.org/oxeanbits/redukt.svg?branch=master)](https://travis-ci.org/oxeanbits/redukt)
+[![](https://jitpack.io/v/oxeanbits/redukt.svg)](https://jitpack.io/#oxeanbits/redukt)
+[![codebeat badge](https://codebeat.co/badges/50fb8d27-6eca-424e-9bbe-6f469b95cec9)](https://codebeat.co/projects/github-com-oxeanbits-redukt-master)
 
 # redukt
 Redux architecture pattern to Android writed in Kotlin
@@ -19,8 +19,8 @@ allprojects {
 Step 2. Add the dependency
 ```gradle
 dependencies {
-    compile 'com.github.raulccabreu.redukt:core:0.1.3'
-    compile 'com.github.raulccabreu.redukt:ui:0.1.3'
+    compile 'com.github.oxeanbits.redukt:core:0.1.3'
+    compile 'com.github.oxeanbits.redukt:ui:0.1.3'
 }
 ```
