+++ core/src/main/AndroidManifest.xml
@@ -1,2 +1,2 @@
-<manifest package="com.github.krysthianlessa.redukt">
+<manifest package="com.github.oxeanbits.redukt">
 </manifest>
