+++ core/src/main/java/com/github/oxeanbits/redukt/actions/Action.kt
@@ -1,0 +1,3 @@
+package com.github.oxeanbits.redukt.actions
+
+data class Action<out T>(val name: String, val payload: T? = null)
