+++ core/src/main/java/com/github/oxeanbits/redukt/actions/Reduce.kt
@@ -1,0 +1,5 @@
+package com.github.oxeanbits.redukt.actions
+
+@Retention(AnnotationRetention.RUNTIME)
+@Target(AnnotationTarget.FUNCTION)
+annotation class Reduce(val action: String)
