+++ core/src/main/java/com/github/oxeanbits/redukt/middlewares/AfterAction.kt
@@ -1,0 +1,5 @@
+package com.github.oxeanbits.redukt.middlewares
+
+@Retention(AnnotationRetention.RUNTIME)
+@Target(AnnotationTarget.FUNCTION)
+annotation class AfterAction(val action: String)
