+++ core/src/main/java/com/github/oxeanbits/redukt/middlewares/AfterActions.kt
@@ -1,0 +1,5 @@
+package com.github.oxeanbits.redukt.middlewares
+
+@Retention(AnnotationRetention.RUNTIME)
+@Target(AnnotationTarget.FUNCTION)
+annotation class AfterActions(val filter: Array<String> = [])
