+++ core/src/main/java/com/github/oxeanbits/redukt/middlewares/Middleware.kt
@@ -1,0 +1,10 @@
+package com.github.oxeanbits.redukt.middlewares
+
+import com.github.oxeanbits.redukt.actions.Action
+
+interface Middleware<in State> {
+
+    fun before(state: State, action: Action<*>)
+
+    fun after(state: State, action: Action<*>)
+}
