+++ core/src/main/java/com/github/oxeanbits/redukt/reducers/Reducer.kt
@@ -1,0 +1,7 @@
+package com.github.oxeanbits.redukt.reducers
+
+import com.github.oxeanbits.redukt.actions.Action
+
+interface Reducer<T> {
+    fun reduce(state: T, action: Action<*>): T
+}
