+++ core/src/main/java/com/github/oxeanbits/redukt/states/StateListener.kt
@@ -1,0 +1,7 @@
+package com.github.oxeanbits.redukt.states
+
+interface  StateListener<in T> {
+    fun hasChanged(newState: T, oldState: T): Boolean
+
+    fun onChanged(state: T)
+}
