+++ core/src/main/java/com/github/oxeanbits/redukt/utils/ConcurrentLinkedHashMapUtils.kt
@@ -1,0 +1,10 @@
+package com.github.oxeanbits.redukt.utils
+
+import com.googlecode.concurrentlinkedhashmap.ConcurrentLinkedHashMap
+
+fun <T> createLinkedMap(): ConcurrentLinkedHashMap<String, T> {
+    return ConcurrentLinkedHashMap
+            .Builder<String, T>()
+            .maximumWeightedCapacity(1000)
+            .build()
+}
