--- core/src/main/java/com/github/raulccabreu/redukt/actions/Action.kt
@@ -1,3 +1,0 @@
-package com.github.raulccabreu.redukt.actions
-
-data class Action<out T>(val name: String, val payload: T? = null)
