--- core/src/main/java/com/github/raulccabreu/redukt/actions/Reduce.kt
@@ -1,5 +1,0 @@
-package com.github.raulccabreu.redukt.actions
-
-@Retention(AnnotationRetention.RUNTIME)
-@Target(AnnotationTarget.FUNCTION)
-annotation class Reduce(val action: String)
