--- core/src/main/java/com/github/raulccabreu/redukt/middlewares/AfterAction.kt
@@ -1,5 +1,0 @@
-package com.github.raulccabreu.redukt.middlewares
-
-@Retention(AnnotationRetention.RUNTIME)
-@Target(AnnotationTarget.FUNCTION)
-annotation class AfterAction(val action: String)
