--- core/src/main/java/com/github/raulccabreu/redukt/reducers/Reducer.kt
@@ -1,7 +1,0 @@
-package com.github.raulccabreu.redukt.reducers
-
-import com.github.raulccabreu.redukt.actions.Action
-
-interface Reducer<T> {
-    fun reduce(state: T, action: Action<*>): T
-}
