--- core/src/main/java/com/github/raulccabreu/redukt/states/StateListener.kt
@@ -1,7 +1,0 @@
-package com.github.raulccabreu.redukt.states
-
-interface  StateListener<in T> {
-    fun hasChanged(newState: T, oldState: T): Boolean
-
-    fun onChanged(state: T)
-}
