--- core/src/main/java/com/github/raulccabreu/redukt/utils/ConcurrentLinkedHashMapUtils.kt
@@ -1,10 +1,0 @@
-package com.github.raulccabreu.redukt.utils
-
-import com.googlecode.concurrentlinkedhashmap.ConcurrentLinkedHashMap
-
-fun <T> createLinkedMap(): ConcurrentLinkedHashMap<String, T> {
-    return ConcurrentLinkedHashMap
-            .Builder<String, T>()
-            .maximumWeightedCapacity(1000)
-            .build()
-}
