+++ core/src/test/java/com/github/oxeanbits/redukt/actions/ActionTest.kt
@@ -1,0 +1,23 @@
+package com.github.oxeanbits.redukt.actions
+
+import org.junit.Assert
+import org.junit.Test
+
+class ActionTest {
+
+    @Test
+    fun whenCreateActionWithoutPayload() {
+        val action = Action<Any?>("TEST")
+
+        Assert.assertNull(action.payload)
+        Assert.assertEquals("TEST", action.name)
+    }
+
+    @Test
+    fun whenCreateActionWithStringPayload() {
+        val action = Action("TEST", "payload")
+
+        Assert.assertEquals("payload", action.payload)
+        Assert.assertEquals("TEST", action.name)
+    }
+}
