+++ ui/src/main/AndroidManifest.xml
@@ -1,2 +1,2 @@
-<manifest package="com.github.raulccabreu.redukt.ui">
+<manifest package="com.github.oxeanbits.redukt.ui">
 </manifest>
