--- ui/src/main/java/com.github.raulccabreu.redukt.ui/ReactiveActivity.kt
@@ -1,24 +1,0 @@
-package com.github.raulccabreu.redukt.ui
-
-import androidx.appcompat.app.AppCompatActivity
-import com.github.raulccabreu.redukt.Redukt
-import com.github.raulccabreu.redukt.states.StateListener
-
-abstract class ReactiveActivity<T> : AppCompatActivity(), StateListener<T> {
-
-    protected abstract fun getRedukt(): Redukt<T>
-
-    override fun onStart() {
-        super.onStart()
-
-        getRedukt().listeners.add(this)
-        onChanged(getRedukt().state)
-    }
-
-    override fun onStop() {
-        getRedukt().listeners.remove(this)
-        super.onStop()
-    }
-
-    override fun hasChanged(newState: T, oldState: T): Boolean = newState != oldState
-}
